Remove logic for creating resource group & DC/OS cluster
