+++ src/main/java/com/microsoft/jenkins/acs/commands/GetPublicFQDNCommand.java
@@ -5,31 +5,37 @@
  */
 package com.microsoft.jenkins.acs.commands;
 
-import com.microsoft.jenkins.acs.exceptions.AzureCloudException;
-import com.microsoft.jenkins.acs.util.NetworkResourceProviderHelper;
-
-import java.io.IOException;
+import com.microsoft.azure.management.Azure;
+import com.microsoft.azure.management.compute.ContainerService;
 
 public class GetPublicFQDNCommand implements ICommand<GetPublicFQDNCommand.IGetPublicFQDNCommandData> {
     @Override
     public void execute(GetPublicFQDNCommand.IGetPublicFQDNCommandData context) {
-        try {
-            context.logStatus("Getting management public FQDN.");
-            String mgmtFQDN = NetworkResourceProviderHelper.getMgmtPublicIPFQDN(
-                    context.getAzureClient(),
-                    context.getResourceGroupName(),
-                    context.getDnsNamePrefix());
-            context.logStatus("Management public FQDN: " + mgmtFQDN);
-            context.setMgmtFQDN(mgmtFQDN);
-            context.setDeploymentState(DeploymentState.Success);
-        } catch (IOException | AzureCloudException e) {
-            context.logError("Error deploying marathon service or enabling ports:", e);
+        context.logStatus("Getting management public FQDN.");
+        final Azure azureClient = context.getAzureClient();
+        final String resourceGroupName = context.getResourceGroupName();
+        final String containerServiceName = context.getContainerServiceName();
+
+        ContainerService containerService = azureClient.containerServices().getByResourceGroup(context.getResourceGroupName(), context.getContainerServiceName());
+        if (containerService == null) {
+            context.logError(String.format("Cannot load Container Service %s from Resource Group %s", containerServiceName, resourceGroupName));
+            return;
         }
+
+        final String fqdn = containerService.masterFqdn();
+        context.logStatus("Management master FQDN: " + fqdn);
+        context.setMgmtFQDN(fqdn);
+
+        final String adminUser = containerService.linuxRootUsername();
+        context.logStatus("Management admin username: " + adminUser);
+        context.setLinuxRootUsername(adminUser);
+
+        context.setDeploymentState(DeploymentState.Success);
     }
 
     public interface IGetPublicFQDNCommandData extends IBaseCommandData {
-        String getDnsNamePrefix();
-
         void setMgmtFQDN(String mgmtFQDN);
+
+        void setLinuxRootUsername(String linuxAdminUsername);
     }
 }
