+++ src/main/java/com/microsoft/jenkins/acs/commands/IBaseCommandData.java
@@ -23,4 +23,6 @@ public interface IBaseCommandData {
     Azure getAzureClient();
 
     String getResourceGroupName();
+
+    String getContainerServiceName();
 }
