+++ src/main/resources/com/microsoft/jenkins/acs/ACSDeploymentContext/config.jelly
@@ -1,47 +1,32 @@
 <?jelly escape-by-default='true'?>
-<j:jelly xmlns:j="jelly:core" xmlns:st="jelly:stapler" xmlns:d="jelly:define" xmlns:l="/lib/layout" xmlns:t="/lib/hudson" xmlns:f="/lib/form">
-    <f:section title="${%ACS_Profile_Configuration}">
-    <f:entry title="${%Location}" field="location" 
-    		help="/plugin/azure-acs-plugin/help-location.html">
-      <f:select default="${descriptor.defaultLocation()}"/>
-    </f:entry>  
-    <f:entry title="${%Dns_Name_Prefix}" field="dnsNamePrefix" 
-             help="/plugin/azure-acs-plugin/help-dnsNamePrefix.html">
-      <f:textbox default="${descriptor.defaultDnsNamePrefix()}" />
+<j:jelly xmlns:j="jelly:core" xmlns:st="jelly:stapler" xmlns:d="jelly:define" xmlns:l="/lib/layout"
+         xmlns:t="/lib/hudson" xmlns:f="/lib/form" xmlns:c="/lib/credentials">
+    <f:entry title="${%azureCredentialsId_title}" field="azureCredentialsId"
+             help="/plugin/azure-acs-plugin/help-azureCredentialsId.html">
+        <c:select expressionAllowed="false"/>
     </f:entry>
-    <f:entry title="${%Agent_Count}" field="agentCount" 
-             help="/plugin/azure-acs-plugin/help-agentCount.html">
-      <f:textbox default="${descriptor.defaultAgentCount()}"/>
+    <f:entry title="${%resourceGroupName_title}" field="resourceGroupName"
+             help="/plugin/azure-acs-plugin/help-resourceGroupName.html">
+        <f:select/>
     </f:entry>
-    <f:entry title="${%Agent_VM_Size}" field="agentVMSize" 
-             help="/plugin/azure-acs-plugin/help-agentVMSize.html">
-        <f:select default="${descriptor.defaultAgentVMSize()}"/>
+
+    <f:entry title="${%containerServiceName_title}" field="containerServiceName"
+             help="/plugin/azure-acs-plugin/help-containerServiceName.html">
+        <f:select/>
     </f:entry>
-    <f:entry title="${%Linux_Admin_Username}" field="linuxAdminUsername" 
-             help="/plugin/azure-acs-plugin/help-linuxAdminUsername.html">
-      <f:textbox default="${descriptor.defaultLinuxAdminname()}"/>
-    </f:entry>
-    <f:entry title="${%Master_Count}" field="masterCount" 
-             help="/plugin/azure-acs-plugin/help-masterCount.html">
-        <f:select />
-    </f:entry>
-    <f:entry title="${%SSH_RSA_Public_Key}" field="sshRSAPublicKey" 
-             help="/plugin/azure-acs-plugin/help-sshRSAPublicKey.html">
-      <f:textbox />
-    </f:entry>  
-  </f:section>
-  <f:section title="${%Marathon_Profile_Configuration}">
-    <f:entry title="${%Marathon_Config_File_Location}" field="marathonConfigFile"
-    	help="/plugin/azure-acs-plugin/help-marathonConfigFile.html">
-      <f:textbox />
-    </f:entry>
-    <f:entry title="${%SSH_RSA_Private_File_Location}" field="sshKeyFileLocation"
-    	help="/plugin/azure-acs-plugin/help-sshKeyFileLocation.html">
-      <f:textbox />
-    </f:entry>
-    <f:entry title="${%SSH_RSA_Private_File_Password}" field="sshKeyFilePassword"
-    	help="/plugin/azure-acs-plugin/help-sshKeyFilePassword.html">
-        <f:password />
-    </f:entry>
-  </f:section>
+
+    <f:section title="${%Marathon_Profile_Configuration}">
+        <f:entry title="${%Marathon_Config_File_Location}" field="marathonConfigFile"
+                 help="/plugin/azure-acs-plugin/help-marathonConfigFile.html">
+            <f:textbox/>
+        </f:entry>
+        <f:entry title="${%SSH_RSA_Private_File_Location}" field="sshKeyFileLocation"
+                 help="/plugin/azure-acs-plugin/help-sshKeyFileLocation.html">
+            <f:textbox/>
+        </f:entry>
+        <f:entry title="${%SSH_RSA_Private_File_Password}" field="sshKeyFilePassword"
+                 help="/plugin/azure-acs-plugin/help-sshKeyFilePassword.html">
+            <f:password/>
+        </f:entry>
+    </f:section>
 </j:jelly>
