+++ src/main/resources/com/microsoft/jenkins/acs/ACSDeploymentContext/config.properties
@@ -1,20 +1,12 @@
-Location=Region
-Max_Virtual_Machines_Limit=Max Virtual Machines Limit	
-Service_Management_URL=Management Service URL
 Verify_Configuration=Verify Configuration
 Verifying=Verifying...
-Azure_Virtual_Machine_Template=Add Azure Virtual Machine Template
-Azure_Virtual_Machine_Template_desc=Azure instances to be provisioned as slaves
-ACS_Profile_Configuration=Azure Container Service Profile Configuration
-Dns_Name_Prefix=DNS Name Prefix
-Agent_Count=Agent Count
-Agent_VM_Size=Agent VM Size
-Linux_Admin_Username=Admin Username
-Orchestrator_Type=Orchestrator Type
-Master_Count=Master Count
-SSH_RSA_Public_Key=SSH RSA Public Key
+
 Marathon_Profile_Configuration=Marathon Profile Configuration
 Marathon_Config_File_Location=Marathon config file path
-Marathon_Config_Host_Port=Host Port
 SSH_RSA_Private_File_Location=SSH RSA private file path
 SSH_RSA_Private_File_Password=SSH RSA private file password
+
+resourceGroupName_title = Resource Group
+containerServiceName_title = Container Service
+
+azureCredentialsId_title = Azure Credentials
