+++ src/main/resources/com/microsoft/jenkins/acs/ACSDeploymentRecorder/config.jelly
@@ -1,10 +1,6 @@
 <?jelly escape-by-default='true'?>
 <j:jelly xmlns:j="jelly:core" xmlns:st="jelly:stapler" xmlns:d="jelly:define" xmlns:l="/lib/layout"
          xmlns:t="/lib/hudson" xmlns:f="/lib/form" xmlns:c="/lib/credentials">
-    <f:entry title="${%azureCredentialsId_title}" field="azureCredentialsId"
-             help="/plugin/azure-acs-plugin/help-azureCredentialsId.html">
-        <c:select expressionAllowed="false"/>
-    </f:entry>
     <f:validateButton title="${%Verify_Configuration}" progress="${%Verifying}" method="verifyConfiguration"
                       with="azureCredentialsId"/>
     <f:property field="context">
