+++ src/main/resources/com/microsoft/jenkins/acs/ACSDeploymentRecorder/config.properties
@@ -23,5 +23,3 @@ Marathon_Config_File_Location=Marathon config file path
 Marathon_Config_Host_Port=Host Port
 SSH_RSA_Private_File_Location=SSH RSA private file path
 SSH_RSA_Private_File_Password=SSH RSA private file password
-
-azureCredentialsId_title = Azure Credentials
