--- src/main/webapp/help-agentCount.html
@@ -1,5 +1,0 @@
-<div>
-	The number of agents for the cluster.  
-	
-	This value can be from 1 to 40 (note, for DC/OS clusters you will also get 1 or 2 public agents in addition to these seleted masters
-</div>
