--- src/main/webapp/help-agentVMSize.html
@@ -1,3 +1,0 @@
-<div>
-	The size of the Virtual Machine.
-</div>
