+++ src/main/webapp/help-containerServiceName.html
@@ -1,0 +1,11 @@
+<div>
+    The name of the container service with Kubernetes orchestrator.
+    <p>
+        Check the following documentation on how to setup DC/OS cluster in Azure Container Service.
+    </p>
+    <ul>
+        <li>
+            <a href="https://docs.microsoft.com/en-us/azure/container-service/container-service-dcos-quickstart" target="_blank" rel="nofollow">Deploy a DC/OS cluster</a>
+        </li>
+    </ul>
+</div>
