--- src/main/webapp/help-dnsNamePrefix.html
@@ -1,5 +1,0 @@
-<div>
-	Sets the Domain name prefix for the cluster.  
-	
-	The concatenation of the domain name and the regionalized DNS zone make up the fully qualified domain name associated with the public IP address.
-</div>
