--- src/main/webapp/help-linuxAdminUsername.html
@@ -1,3 +1,0 @@
-<div>
-	User name for the Linux Virtual Machines.
-</div>
