--- src/main/webapp/help-location.html
@@ -1,3 +1,0 @@
-<div>
-	Azure regions.
-</div>
