--- src/main/webapp/help-masterCount.html
@@ -1,3 +1,0 @@
-<div>
-	The number of DC/OS masters for the cluster.
-</div>
