+++ src/main/webapp/help-resourceGroupName.html
@@ -1,0 +1,3 @@
+<div>
+    Resource group of the Kubernetes cluster container service.
+</div>
