--- src/main/webapp/help-sshRSAPublicKey.html
@@ -1,5 +1,0 @@
-<div>
-	Configure all linux machines with the SSH RSA public key string.  
-	
-	Your key should include three parts, for example 'ssh-rsa AAAAB...snip...UcyupgH azureuser@linuxvm'
-</div>
